move one test; separate some files
