+++ irene-core/src/test/kotlin/edu/umass/cics/ciir/irene/indexing/LDocBuilderTest.kt
@@ -1,7 +1,6 @@
 package edu.umass.cics.ciir.irene.indexing
 
 import edu.umass.cics.ciir.irene.utils.forAllPairs
-import edu.umass.cics.ciir.irene.IndexParams
 import edu.umass.cics.ciir.irene.lang.*
 import edu.umass.cics.ciir.irene.tokenize
 import org.junit.Assert.assertEquals
