+++ irene-galago/src/main/kotlin/edu/umass/cics/ciir/irene/collections/IndexRobust2004.kt
@@ -1,6 +1,6 @@
 package edu.umass.cics.ciir.irene.collections
 
-import edu.umass.cics.ciir.irene.IndexParams
+import edu.umass.cics.ciir.irene.indexing.IndexParams
 import edu.umass.cics.ciir.irene.indexing.IreneIndexer
 import edu.umass.cics.ciir.irene.docs
 import edu.umass.cics.ciir.irene.utils.CountingDebouncer
